LayoutTests/imported/w3c/web-platform-tests/domxpath/resolver-callback-interface-cross-realm-expected.txt

CONSOLE MESSAGE: TypeError: 'lookupNamespaceURI' property of callback interface should be callable
CONSOLE MESSAGE: TypeError: 'lookupNamespaceURI' property of callback interface should be callable
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it


PASS XPathNSResolver is cross-realm plain object without 'lookupNamespaceURI' property
PASS XPathNSResolver is cross-realm plain object with non-callable 'lookupNamespaceURI' property
FAIL XPathNSResolver is cross-realm plain object with revoked Proxy as 'lookupNamespaceURI' property promise_rejects_js: function "function () { throw e }" threw object "TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it" ("TypeError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS XPathNSResolver is cross-realm non-callable revoked Proxy
PASS XPathNSResolver is cross-realm callable revoked Proxy
